requirements.txt
# requirements.txt
# Core scientific computing
numpy>=1.19.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
xgboost>=1.5.0
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0

# Data handling and I/O
uproot>=4.0.0
awkward>=1.5.0
h5py>=3.0.0
zarr>=2.8.0
dask[dataframe]>=2021.0
dask[distributed]>=2021.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress and utilities
tqdm>=4.62.0
joblib>=1.1.0
psutil>=5.8.0
pyyaml>=5.4.0
click>=8.0.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.0.0
black>=21.0.0
isort>=5.0.0
flake8>=3.8.0

# Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=7.0.0
